[refactor] database form 분리
